refactor: rename cli.commands module to cli.repl_commands
Free the `akgentic.infra.cli.commands` name for a new top-level Typer
command subpackage (forthcoming login/logout commands). Python does not
permit a module and a subpackage with the same name at the same level,
so the legacy in-session REPL slash-command registry moves here. No
behavioral change — all import sites in src/ and tests/ are updated.

Relates to #235
