+++ src/akgentic/infra/cli/tui/app.py
@@ -29,7 +29,7 @@ _log = logging.getLogger(__name__)
 
 if TYPE_CHECKING:
     from akgentic.infra.cli.client import ApiClient
-    from akgentic.infra.cli.commands import CommandRegistry
+    from akgentic.infra.cli.repl_commands import CommandRegistry
     from akgentic.infra.cli.connection import ConnectionManager
     from akgentic.infra.cli.event_router import EventRouter
