+++ src/akgentic/infra/cli/tui/command_adapter.py
@@ -9,7 +9,7 @@ from contextlib import redirect_stdout
 from typing import TYPE_CHECKING
 
 from akgentic.infra.cli.client import ApiClient, ApiError
-from akgentic.infra.cli.commands import CommandRegistry
+from akgentic.infra.cli.repl_commands import CommandRegistry
 from akgentic.infra.cli.renderers import RichRenderer
 
 if TYPE_CHECKING:
