+++ tests/cli/test_cli_commands_insession.py
@@ -18,7 +18,7 @@ from akgentic.infra.cli.client import (
     WorkspaceTreeInfo,
     WorkspaceUploadInfo,
 )
-from akgentic.infra.cli.commands import (
+from akgentic.infra.cli.repl_commands import (
     CommandRegistry,
     _agents_handler,
     _catalog_handler,
@@ -301,7 +301,7 @@ class TestDeleteHandler:
         client = _mock_client()
         session = _make_session(client=client)
 
-        with patch("akgentic.infra.cli.commands.builtins.input", return_value="y"):
+        with patch("akgentic.infra.cli.repl_commands.builtins.input", return_value="y"):
             await _delete_handler("", session)
 
         out = capsys.readouterr().out
@@ -312,7 +312,7 @@ class TestDeleteHandler:
         client = _mock_client()
         session = _make_session(client=client)
 
-        with patch("akgentic.infra.cli.commands.builtins.input", return_value="n"):
+        with patch("akgentic.infra.cli.repl_commands.builtins.input", return_value="n"):
             await _delete_handler("", session)
 
         out = capsys.readouterr().out
@@ -325,7 +325,7 @@ class TestDeleteHandler:
         client = _mock_client()
         session = _make_session(client=client)
 
-        with patch("akgentic.infra.cli.commands.builtins.input", return_value="y"):
+        with patch("akgentic.infra.cli.repl_commands.builtins.input", return_value="y"):
             await _delete_handler("", session)
 
         client.get_team.assert_called_once_with("t1")
@@ -337,7 +337,7 @@ class TestDeleteHandler:
         client = _mock_client()
         session = _make_session(client=client)
 
-        with patch("akgentic.infra.cli.commands.builtins.input", return_value="y"):
+        with patch("akgentic.infra.cli.repl_commands.builtins.input", return_value="y"):
             await _delete_handler("", session)
 
         out = capsys.readouterr().out
